pyproject.toml
[project]
name = "inventoryagent"
version = "0.1.0"
description = "Inventory management system for kit based products"
authors = [
    {name = "quent-dev quentin.devacct@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth-oauthlib (>=1.2.3,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "streamlit (>=1.52.2,<2.0.0)",
    "plotly (>=6.5.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
